backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.2.1",
]
